mcb_domain::ports::validation

Trait Validator

pub trait Validator: Send + Sync {
    // Required method
    fn name(&self) -> &'static str;

    // Provided methods
    fn checks<'a>(
        &'a self,
        _config: &'a ValidationConfig,
    ) -> ValidatorResult<Vec<NamedCheck<'a>>> { ... }
    fn validate(
        &self,
        config: &ValidationConfig,
    ) -> ValidatorResult<Vec<Box<dyn Violation>>> { ... }
    fn enabled_by_default(&self) -> bool { ... }
    fn description(&self) -> &'static str { ... }
    fn supported_languages(&self) -> &[LanguageId] { ... }
}

Interface for implementing codebase validation rules.

Validators are responsible for checking certain aspects of the code (naming, SOLID, cleanliness) and returning a list of violations found.

Required Methods

fn name(&self) -> &'static str

Returns the unique name of this validator.

Provided Methods

fn checks<'a>( &'a self, _config: &'a ValidationConfig, ) -> ValidatorResult<Vec<NamedCheck<'a>>>

Returns a list of individual named checks for this validator.

Errors

Returns a ValidatorResult which may contain an error if check preparation fails.

fn validate( &self, config: &ValidationConfig, ) -> ValidatorResult<Vec<Box<dyn Violation>>>

Performs the full validation scan for this validator.

Errors

Returns a ValidatorResult error if the validation process fails.

fn enabled_by_default(&self) -> bool

Check if this validator should run by default in a standard scan.

fn description(&self) -> &'static str

Get a human-readable description of what this validator checks.

fn supported_languages(&self) -> &[LanguageId]

Get the list of languages supported by this validator.